streamlit_ui/requirements.txt
streamlit>=1.24
torch>=1.12
torchvision
torchaudio
pillow
librosa
numpy
pandas
